requirements.txt
crewai>=0.28.0
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-community>=0.0.13
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0
openai>=1.3.0
tiktoken>=0.5.1
pydantic>=2.4.2
pyyaml>=6.0.1 